Device notify (#6)
Removed a direct relationship between our sample frontend app and Azure Notification Hub. The goal of this change is to abstract device management to the backend in order to support multiple devices per user. The envisioned scenario is that a user might have a table, phone, and laptop. During a user session they change tags on their phone; if we have a direct connection to Notification Hub we can't update tags for the other two devices. However, by abstracting it to an API we are able to maintain user devices and update all devices at the same time. This initial commit allows us to go down that direction.

The next phase will see us introduce a user-device database that will track all installations that belong to a user and allow shared device updates.