Function for saving devices to database
Created a function that takes in an input with basic device details
including the handle, platform, and tags.  That is then translated to a
database model which includes other details such as time stamps.
Ultimately this data will be used to send notifications to devices using
Notification Hub purely as a passthrough service that does not store
device information.  Device information is handed off to an external
database, which in this case is CosmosDB.

[request] --> f(upsert-device) --> CosmosDB

Using an external database can simplify device management by creating a
schema with 1..M relationship between [Users]..[Devices]. Ultimately
ending in a structure where Tags are associated to a [User] not a device

When a notification is ready for sent any field in the data model can be
queried to retrieve the correct device and handle. Finally with the
handle(s) retrieved you can use Notification Hub to directly push that
notification to the device(s)