Terraform definitions (#18)
* Notification hub definition

Added terraform defining the resource group, namespace, and default hub
for push notifications. Extracted configurations out into terraform
variables to simplify setup and reduce hardcoding for reusability.

* GCM Credential dynamic block

Made the `gcm_credential` block dynamic based on the value of the
variable `gcm_api_key`.  If a value is provided to the variable then the
block will be set.

* Changed the dynamic block to use splat expression

* Removed unused locals

* Terraform definitions

Added the terraform definitions needed to deploy the notification hub
instance of the system.  This defines the reosource group, notification
hub, and functions app resources needed for the demo to work

* Terraform AppSettings improvements

Added terraform blocks and elements that handle function app
configuration. Terrafrm will now properly configure and set the
environment variables `NOTIFICAITON_HUB_NAME` and `NOTIFICATION_HUB_CS`.
Both of these are required to properly initialize the hub client.

* Fixed ci-cd yaml to include updated and consistent infra names