Event handler function
Added a function to handle EventGrid messages and persist the email
event to database. Data is written to two seperate locations (akin to
tables).  One location is for the sent email message, and the second
location is for events that correspond to a message. Events include
delivery notifications (success, fail) and engagement tracking. This
allows users to perform analytics based on anonymous engagment and
delivery success rates.