Readme and cleanup
Started readme to better explain the ACS architecture and system. It
also details how to quickly deploy a demo system using the deployment
script under `/scripts/deployments/acs`. Added an ingress function to
support indirect message sending.  This ingress function is just a
frontend to the queue that triggers all the work. However, adding the
function better aligns with the architectural concept of loose coupling.