Azure Communication Services Email (#29)
* Adding Azure Communication Services

This adds an example of using Azure Communication Services email and how
to utilize the service to meet the following criteria:

- Email sending
- Email status
- Engagement tracking

To enable status tracking we will have to integrate with event grid

* Email Queue Processor

Added function that triggers on a new queue message and sends an email
via Azure Communication Services - Email.  Upon delivering the message
it saves the message Id in a backend store for status and engagement
tracking.

* Email persistence

Added data store to track message send requests and events associated
with email messages. Such as status change, engagement, etc.

* Event handler function

Added a function to handle EventGrid messages and persist the email
event to database. Data is written to two separate locations (akin to
tables).  One location is for the sent email message, and the second
location is for events that correspond to a message. Events include
delivery notifications (success, fail) and engagement tracking. This
allows users to perform analytics based on anonymous engagement and
delivery success rates.

* Terraform ACS definitions

Added initial definitions to provision Azure Communication Services
using the Terraform provider Azure/azapi.  This provider a thin layer
that interfaces with the ARM API.  The hashi/azurerm provider does not
have definitions to ACS so the script is using a combination of both
providers to accomplish its task

* Adding terraform definitions for the acs architecture flow

* Terraform upgrades

Added several new modules to better segment our IaC. This will also
allow us to add new demos in the future reusing existing modules, or
expanding on existing modules.  This nearly completes the ACS IaC
script. The only remaining item is to figure out how to create logic-app
connections through IaC (if recommended and possible)

* Terraform and automation

Finalizing the terraform to generate the Azure Communication Services
system architecture in one go.  This will allow users to quickly spin up
an ACS instance and all other resources to send, track, and monitor
engagement on emails.  I am also adding a powershell script that allows
users to do all this from their local dev environment for testing.

* Function App Deployment

Fixed the powershell script that handles function app deployment using
zip file.  I wasn't properly configuring the OS and the zip instructions
for the app.  Making a few powershell changes fixed the deployment

* Made function app name a parameter

Made function app name a parameter to allow for more customization

* LogicApp defintion + Property promotions grid

Added the initial logic app definition for deployment. Added support to
promote properties in event grid subscriptions.

* Added some steps to logic app for testing

* Breaking apart ignore files for easier management

* Readme and cleanup

Started readme to better explain the ACS architecture and system. It
also details how to quickly deploy a demo system using the deployment
script under `/scripts/deployments/acs`. Added an ingress function to
support indirect message sending.  This ingress function is just a
frontend to the queue that triggers all the work. However, adding the
function better aligns with the architectural concept of loose coupling.

* Readme udpate

Added more details on the setup and what nuget package was used for ACS
sdk.

* Added tfvars

Added a vars file for terraform to make modification easier

* Read me updates

Adding more details to the readme file so users can be up and running
faster

* Updated arch diagram

* Model validation

Added validation to input models for HTTP and SB triggers.  Now when a
request is submitted the model state is validated to ensure required
fields are populated. Later on we will add more validation attributes to
enforce property values

* Updated structure log message format

* fixed arch diagram and updated readme

* Removed files that shouldn't have been commited

* removed unused connection